future proof

+++ contracts/NerwoEscrow.sol
@@ -52,7 +52,7 @@ contract NerwoEscrow is Ownable, Initializable, ReentrancyGuard {
 
     struct Transaction {
         Status status;
-        uint32 lastInteraction; // Last interaction for the dispute procedure.
+        uint64 lastInteraction; // Last interaction for the dispute procedure.
         address client;
         address freelancer;
         IERC20 token;
@@ -72,7 +72,7 @@ contract NerwoEscrow is Ownable, Initializable, ReentrancyGuard {
     mapping(IERC20 => bool) public tokens; // whitelisted ERC20 tokens
 
     struct ArbitratorData {
-        uint32 feeTimeout; // Time in seconds a party can take to pay arbitration
+        uint64 feeTimeout; // Time in seconds a party can take to pay arbitration
         // fees before being considered unresponding and lose the dispute.
         IArbitrator arbitrator; // Address of the arbitrator contract.
         IArbitrableProxy proxy; // Address of the arbitrator proxy contract.
@@ -221,7 +221,7 @@ contract NerwoEscrow is Ownable, Initializable, ReentrancyGuard {
         address arbitratorProxy,
         bytes calldata arbitratorExtraData
     ) internal {
-        arbitratorData.feeTimeout = uint32(feeTimeout);
+        arbitratorData.feeTimeout = uint64(feeTimeout);
         arbitratorData.arbitrator = IArbitrator(arbitrator);
         arbitratorData.proxy = IArbitrableProxy(arbitratorProxy);
         arbitratorData.extraData = arbitratorExtraData;
@@ -365,7 +365,7 @@ contract NerwoEscrow is Ownable, Initializable, ReentrancyGuard {
 
         _transactions[transactionID] = Transaction({
             status: Status.NoDispute,
-            lastInteraction: uint32(block.timestamp),
+            lastInteraction: uint64(block.timestamp),
             client: client,
             freelancer: freelancer,
             token: token,
@@ -462,7 +462,7 @@ contract NerwoEscrow is Ownable, Initializable, ReentrancyGuard {
             revert InvalidAmount();
         }
 
-        transaction.lastInteraction = uint32(block.timestamp);
+        transaction.lastInteraction = uint64(block.timestamp);
 
         if (sender == transaction.client) {
             if (transaction.clientFee != 0) {
